% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeCDR3.R
\docType{methods}
\name{analyzeCDR3}
\alias{analyzeCDR3}
\alias{analyzeCDR3,IMGThelper-method}
\title{description of function analyzeCDR3}
\usage{
analyzeCDR3(x, fname = "6_Junction.txt", OK = NULL)
}
\arguments{
\item{x}{the IMGThelper object}

\item{fname}{the absolute position of the Junction file to analyze ( default( file.path(x$path, 6_Junction.txt')))}

\item{OK}{which rows to process (default which(data$V.DOMAIN.Functionality == 'productive' & ! data$D.GENE.and.allele == "" & ! data$CDR3.IMGT == "" & ! data$CDR3.IMGT == "NA"))}
}
\description{
Stijn - please help here!
}
