% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\docType{methods}
\name{compare}
\alias{compare}
\alias{compare,IMGThelper-method}
\title{description of function compare}
\usage{
compare(x, a)
}
\arguments{
\item{x}{one IMGThelper object}

\item{a}{the other IMGThelper object}
}
\description{
comapre two IMGT result datasets
}
