% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001_class.R
\docType{class}
\name{IMGThelper}
\alias{IMGThelper}
\title{Class a simple interface to biological data (numeric) and rich annotation for both columns (samples) and rows (values)}
\format{\code{\link{R6Class}} object.}
\usage{
IMGThelper
}
\value{
Object of \code{\link{R6Class}} to access IMGT data in R
}
\description{
Class a simple interface to biological data (numeric) and rich annotation for both columns (samples) and rows (values)
}
\section{Fields}{

\describe{
\item{\code{path}}{the path to the IMGT data}

\item{\code{cells}}{the cells table refering to the detected cells analyzed in this 10x VDJ experiment}

\item{\code{usedObj}}{a multi purpose list to store whichever ananlyis results do not fit in the stats list}

\item{\code{version}}{the version of the IMGThelper package}
}}

\keyword{IMGThelper}
