% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FreqNoNUC.R
\docType{methods}
\name{FreqNoNNuc}
\alias{FreqNoNNuc}
\alias{NoNNuc,IMGThelper-method}
\title{description of function NoNNuc}
\usage{
FreqNoNNuc(x, minCount = 0)
}
\arguments{
\item{x}{the IMGThelper object}

\item{minCount}{the minimal occurance of one VDJ element ( default = 0)}
}
\description{
calculate the frequency of VDJ segments not containing N nucleotides
}
