% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\docType{methods}
\name{open}
\alias{open}
\alias{open,IMGThelper-method}
\title{description of function open}
\usage{
open(x, name = "6_Junction.txt")
}
\arguments{
\item{x}{the IMGThelper object}

\item{name}{the file to load default= '6_Junction.txt'}
}
\description{
load one file in the path of the object
}
