% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productive.R
\docType{methods}
\name{productive}
\alias{productive}
\alias{productive,IMGThelper-method}
\title{select only productively recombined VDJ segments}
\usage{
productive(x, fname = file.path(x$path, "6_Junction.txt"),
  OK = function(data) {     which(t(data[, "V.DOMAIN.Functionality"]) ==
  "productive" & !t(data[, "D.GENE.and.allele"]) == "" & !t(data[,
  "CDR3.IMGT"]) == "" & !t(data[, "CDR3.IMGT"]) == "NA") })
}
\arguments{
\item{x}{the IMGThelper object}

\item{fname}{the file to read from default=file.path(x$path, '6_Junction.txt')}

\item{OK}{the ids to select (default = function(data) { which(data$V.DOMAIN.Functionality == 'productive' & ! data$D.GENE.and.allele == "" & ! data$CDR3.IMGT == "" & ! data$CDR3.IMGT == "NA")} )}
}
\description{
select productively recombined VDJ recombination evens from the Junctions file
}
