% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacked_bargraph.R
\docType{methods}
\name{stacked_bargraph}
\alias{stacked_bargraph}
\alias{stacked_bargraph,IMGThelper-method}
\title{description of function stacked_bargraph}
\usage{
stacked_bargraph(x, ofile, name)
}
\arguments{
\item{x}{the IMGThelper object}

\item{ofile}{the outfile (pdf)}

\item{name}{the name of this sample}
}
\description{
create a stacked_bargraph based on the 6_Juntions.txt file
}
