% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\docType{methods}
\name{treemap}
\alias{treemap}
\alias{treemap,IMGThelper-method}
\title{description of function treemap}
\usage{
treemap(x, name, ofile)
}
\arguments{
\item{x}{the IMGThelper object}

\item{name}{the experiment this VDJ analysis came from}

\item{ofile}{the pdf file name}
}
\description{
draw a treemap of clone sizes
}
