% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDJusage.R
\docType{methods}
\name{VDJusage}
\alias{VDJusage}
\alias{VDJusage,IMGThelper-method}
\title{description of function VDJusage}
\usage{
VDJusage(x, minCount = 0)
}
\arguments{
\item{x}{the IMGThelper object}

\item{minCount}{the minimal reoccurance value for a single VDJ recombination event (default = 0)}
}
\description{
returns the cout for each V, D and J segment in the recurringProductives
}
