% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VJcombination.R
\docType{methods}
\name{VJcombination}
\alias{VJcombination}
\alias{VJcombination,IMGThelper-method}
\title{Ask Stijn about this function!}
\usage{
VJcombination(x, cells = NULL, cutoff = 5)
}
\arguments{
\item{x}{the IMGThelper object}

\item{cells}{the cell IDs to look into (default NULL == all)}

\item{cutoff}{select only CDR3 elements used more than cutoff times (default=5; see recurringProductives)}
}
\description{
Function to create VJ_usage.csv files used for the Chord diagrams
}
