% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChordDiagram.R
\docType{methods}
\name{ChordDiagram}
\alias{ChordDiagram}
\alias{ChordDiagram,IMGThelper-method}
\title{description of function ChordDiagram}
\usage{
ChordDiagram(x, ofile, col = NULL)
}
\arguments{
\item{x}{the IMGThelper object}

\item{ofile}{the pdf out file}

\item{col}{the colors to be used in the plot (leave NULL if unsure!)}
}
\description{
Create a Chord diagram of V gene usage
}
