% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\docType{methods}
\name{tree}
\alias{tree}
\alias{tree,IMGThelper-method}
\title{description of function tree}
\usage{
tree(x)
}
\arguments{
\item{x}{the IMGThelper object}
}
\description{
create the data table used for the treemap plots (based on 6_Junctions.txt)
}
