% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VJcombination.R
\docType{methods}
\name{VJcombination}
\alias{VJcombination}
\alias{VJcombination,IMGThelper-method}
\title{Ask Stijn about this function!}
\usage{
VJcombination(x, fname = file.path(x$path, "6_Junction.txt"))
}
\arguments{
\item{x}{the IMGThelper object}

\item{fname}{the file to parse default=file.path(x$path, '6_Junction.txt')}
}
\description{
Function to create VJ_usage.csv files used for the Chord diagrams
}
