% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VJcombination.R
\docType{methods}
\name{VJcombination}
\alias{VJcombination}
\alias{VJcombination,IMGThelper-method}
\title{Ask Stijn about this function!}
\usage{
VJcombination(x, splitAt = 5)
}
\arguments{
\item{x}{the IMGThelper object}

\item{splitAt}{separate the plot at splitAt (ask Stijn of how to phrase that) default= 5}
}
\description{
Function to create VJ_usage.csv files used for the Chord diagrams
}
